Fix calendar size shown as 4GB / never refetched on 64-bit
PVS-Studio V560: CPropCalendarGeneral compared a uint32_t calendar
size against ULONG_MAX. CCalendarStoreNode::GetSize() returns uint32_t
and the "size unknown" sentinel is UINT32_MAX (mSize = UINT32_MAX in
all three ctors). On LP64 ULONG_MAX is 0xFFFFFFFFFFFFFFFF, so the
unknown-size sentinel can never match: the offline/web-calendar
CheckSize() refetch guard (size == ULONG_MAX) never fires, and the
display guard (size != ULONG_MAX) is always true, so an unknown size
is shown as 4294967295 bytes instead of being refetched/blank.

Use uint32_t for the local and compare against UINT32_MAX, matching
GetSize()'s return type and the already-correct multi-calendar block
in the same files. Completes the partial sentinel fix from 963bc461,
which corrected the field init and the multi-calendar comparison but
missed the single-node block (Linux) and both blocks (MacOS/Win32).
Latent since the 2009 original (f939068) where it was 32-bit-correct
(ULONG_MAX == 0xFFFFFFFF on ILP32).

Fixed on all three platforms; PVS scans Linux only.

+++ MacOS/Sources/Application/Calendar/Calendar_Properties/CPropCalendarGeneral.cpp
@@ -124,13 +124,13 @@ void CPropCalendarGeneral::SetCalendar(calstore::CCalendarStoreNode* node)
 	if (!node->IsDirectory())
 	{
 		{
-			unsigned long size = node->GetSize();
-			if ((size == ULONG_MAX) && (node->GetProtocol()->IsOffline() || node->GetProtocol()->IsWebCalendar()))
+			uint32_t size = node->GetSize();
+			if ((size == UINT32_MAX) && (node->GetProtocol()->IsOffline() || node->GetProtocol()->IsWebCalendar()))
 			{
 				node->CheckSize();
 				size = node->GetSize();
 			}
-			if (size != ULONG_MAX)
+			if (size != UINT32_MAX)
 			{
 				mSize->SetText(::GetNumericFormat(size));
 			}
@@ -217,7 +217,7 @@ void CPropCalendarGeneral::CheckSize()
 
 	{
 		uint32_t size = mCalList->front()->GetSize();
-		if (size != ULONG_MAX)
+		if (size != UINT32_MAX)
 		{
 			mSize->SetText(::GetNumericFormat(size));
 		}
